--- a/pyproject.toml
+++ b/pyproject.toml
@@ -7,6 +7,7 @@ requires-python = ">=3.12"
 dependencies = [
     "lxml>=6.0.0",
     "pandas>=2.3.0",
+    "polars>=1.33.1",
     "psycopg2-binary>=2.9.10",
     "sqlalchemy>=2.0.41",
     "tqdm>=4.67.1",