hitaus harmitti ku testaili github actionsei, ni tiputetaas vaski parseri 3min -> 10s

ekaks profiloidaan mistä hitaus tulee, `line_profiler` on kiva ja asentuu `uv add line_profiler`

sit profiloidaan ite skribula `uv run kernprof -l -v pipes/vaski_parser.py`:

```
Timer unit: 1e-06 s

Total time: 185.304 s
File: pipes/vaski_parser.py
Function: vaski_parser at line 9

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
     9                                           @profile
    10                                           def vaski_parser():
    11         1         81.1     81.1      0.0      pd.options.mode.copy_on_write = True  # Silence a warning about pandas. See https://pandas.pydata.org/pandas-docs/stable/user_guide/indexing.html#returning-a-view-versus-a-copy
    12                                           
    13         1         15.5     15.5      0.0      vaski_fname = os.path.join("data", "raw", "VaskiData.tsv")
    14                                           
    15        32   73644115.8  2.3e+06     39.7      for vaski_data in tqdm(pd.read_csv(vaski_fname, sep="\t", chunksize=10000), total=31):  # Total achieved by experiment
    16        62   59057615.8 952542.2     31.9          vaski_data['Xml'] = vaski_data["XmlData"].apply(
    17        31         37.9      1.2      0.0              lambda x: etree.fromstring(x)
    18                                                   )
    19        62      57218.3    922.9      0.0          vaski_data = vaski_data[  # Only keep finnish files
    20        62     471448.0   7604.0      0.3              vaski_data['Xml'].apply(
    21        31         28.8      0.9      0.0                  lambda x: x[0][0].text.endswith('_fi')
    22                                                       )
    23                                                   ]
    24        62     341058.9   5500.9      0.2          vaski_data['doctype'] = vaski_data['Xml'].apply(
    25        31         28.2      0.9      0.0              lambda x: x[0][0].text[13:]  # All doctypes start with VASKI_JULKVP_
    26                                                   )
    27                                           
    28      1423      17664.2     12.4      0.0          for doctype in vaski_data['doctype'].unique():
    29      2784    3080048.6   1106.3      1.7              sub_vaski = vaski_data[vaski_data['doctype'] == doctype].drop(
    30      1392       1062.6      0.8      0.0                  columns=["doctype", 'Xml']
    31                                                       )
    32      1392      20919.4     15.0      0.0              f_path = os.path.join("data", "raw", "vaski", f"{doctype}.tsv")            
    33      1392      21747.1     15.6      0.0              if not os.path.exists(f_path):  # If no file, write header
    34                                                           sub_vaski.to_csv(f_path, sep="\t", header=True, index=False)
    35                                                       else:  # If file, no header and append
    36      1392   48591207.5  34907.5     26.2                  sub_vaski.to_csv(f_path, sep="\t", header=False, index=False, mode='a')
```

kesto 180s, 30% ajasta menee xmlän parsimiseen, mut me ei tehä sil oikeestaa mitää ja voidaa vaa käyttää nopeit regexei tsekkaa doctype ja kieli.

```
Timer unit: 1e-06 s

Total time: 110.598 s
File: pipes/vaski_parser.py
Function: vaski_parser at line 9

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
     9                                           @profile
    10                                           def vaski_parser():
    11         1         87.6     87.6      0.0      pd.options.mode.copy_on_write = True  # Silence a warning about pandas. See https://pandas.pydata.org/pandas-docs/stable/user_guide/indexing.html#returning-a-view-versus-a-copy
    12                                           
    13         1         16.5     16.5      0.0      vaski_fname = os.path.join("data", "raw", "VaskiData.tsv")
    14                                           
    15        32   56743866.1 1.77e+06     51.3      for vaski_data in tqdm(pd.read_csv(vaski_fname, sep="\t", chunksize=10000), total=31):  # Total achieved by experiment
    16        31    1019717.7  32894.1      0.9          vaski_data['doctype'] = vaski_data["XmlData"].str.extract(r"(VASKI_JULKVP_\w+)")
    17        31     302458.1   9756.7      0.3          vaski_data = vaski_data[vaski_data['doctype'].str.endswith('_fi')]
    18                                           
    19      1423      18644.5     13.1      0.0          for doctype in vaski_data['doctype'].unique():
    20      1392    3151755.2   2264.2      2.8              sub_vaski = vaski_data[vaski_data['doctype'] == doctype].drop(columns=["doctype"])
    21      1392      22631.5     16.3      0.0              f_path = os.path.join("data", "raw", "vaski", f"{doctype}.tsv")            
    22      1392      22584.9     16.2      0.0              if not os.path.exists(f_path):  # If no file, write header
    23        71    3027172.2  42636.2      2.7                  sub_vaski.to_csv(f_path, sep="\t", header=True, index=False)
    24                                                       else:  # If file, no header and append
    25      1321   46288838.0  35040.8     41.9                  sub_vaski.to_csv(f_path, sep="\t", header=False, index=False, mode='a')
```
180s -> 110s
helpot 60s pois, mut nyt pullonkaulana on pandasin csv luku ja kirjotus, oiskohan polars kiva

```
Wrote profile results to 'vaski_parser.py.lprof'
Timer unit: 1e-06 s

Total time: 6.87593 s
File: pipes/vaski_parser.py
Function: vaski_parser at line 10

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
    10                                           @profile
    11                                           def vaski_parser():
    12         1         25.9     25.9      0.0      vaski_fname = os.path.join("data", "raw", "VaskiData.tsv")
    13                                           
    14        32    2035592.2  63612.3     29.6      for vaski_data in tqdm(pl.read_csv(vaski_fname, separator="\t", has_header=True).iter_slices(n_rows=10_000), total=31):  # Total achieved by experiment
    15        62     105718.6   1705.1      1.5          vaski_data = vaski_data.with_columns(
    16        31       2639.7     85.2      0.0              pl.col("XmlData").str.extract(r"VASKI_JULKVP_(\w+)").alias("doctype")
    17                                                   )
    18        31      23743.5    765.9      0.3          vaski_data = vaski_data.filter(pl.col("doctype").str.ends_with("_fi"))
    19                                           
    20      1423     298467.2    209.7      4.3          for (doctype,), sub_vaski in vaski_data.group_by("doctype"):
    21      1392      25493.5     18.3      0.4              f_path = os.path.join("data", "raw", "vaski", f"{doctype}.tsv")            
    22      1392     378903.9    272.2      5.5              sub_vaski = sub_vaski.drop("doctype")
    23      2784      78876.2     28.3      1.1              with open(f_path, mode="a" if os.path.exists(f_path) else "w") as f:
    24      1392    3926473.8   2820.7     57.1                  sub_vaski.write_csv(f, separator="\t", include_header=not os.path.exists(f_path))
```

110s -> 7s
nice, ilmast perffii

paha sanoo miks pandas on noi paljo hitaampi mut sitä en jaksanu alkaa selvittää, jätetään lukijalle tehtäväks